requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dateutil==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==23.2.1
boto3==1.35.0
botocore==1.35.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
